Camera Follow
Added smooth follow. But it doesn't keep up with direction changes. Also
added a floating script I was experimenting with a while back. Updated
the mobile joystick stuff to fix a bug where it was conflicting with
camera moves (this is the first 3D project I tried it in. General
project cleanup/hierarchy naming.